pyproject.toml
[tool.poetry]
name = "github-actions-practice"
version = "0.1.0"
description = "GitHub Actions practice repository"
authors = ["Armen <your-email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
flask = "^3.0.0"
schedule = "^1.2.0"
appium-python-client = "^5.1.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
